+++ packages/expo-router/src/native-tabs/utils/icon.ts
@@ -7,6 +7,8 @@ import type {
 } from 'react-native-screens';
 import type { SFSymbol } from 'sf-symbols-typescript';
 
+import { isChildOfType } from '../../utils/children';
+import { NativeTabsTriggerVectorIcon } from '../common/elements';
 import type { NativeTabOptions, NativeTabsProps } from '../types';
 
 export function convertIconColorPropToObject(iconColor: NativeTabsProps['iconColor']): {
@@ -99,6 +101,16 @@ export function convertOptionsIconToAndroidPropsIcon(
   }
   if (icon && 'src' in icon && icon.src) {
     return { type: 'imageSource', imageSource: icon.src };
+  }
+  return undefined;
+}
+
+export function convertComponentSrcToImageSource(src: React.ReactElement) {
+  if (isChildOfType(src, NativeTabsTriggerVectorIcon)) {
+    const props = src.props;
+    return { src: props.family.getImageSource(props.name, 24, 'white') };
+  } else {
+    console.warn('Only VectorIcon is supported as a React element in Icon.src');
   }
   return undefined;
 }